% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beers.R
\docType{data}
\name{beers}
\alias{beers}
\title{A dataframe of information on canned beer.}
\format{A data frame with 2410 rows and 10 variables.}
\usage{
beers
}
\description{
\itemize{
  \item ID \code{character}. Database entry number.
  \item Beer \code{character}. Beer name.
  \item Style \code{character}. Style of the beer.
  \item Ounces \code{character}. Full names of Brazilian states.
  \item Litres \code{character}. Full names of Brazilian states.
  \item ABV_percent \code{character}. Full names of Brazilian states.
  \item IBUs \code{character}. Full names of Brazilian states.
  \item Brewery \code{character}. Brewery name.
  \item City \code{character}. City where the brewery is.
  \item State \code{character}. State where the brewery is located.
}
}
\keyword{datasets}
