% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beer_table.R
\name{beer_table}
\alias{beer_table}
\title{beer_table}
\usage{
\code{beer_table()}
}
\description{
Returns a dataframe of information on canned beer from CraftCans.com.
}
\seealso{
\href{http://www.craftcans.com/about.php}{Craft Cans}
}
